requirements.txt
annotated-types==0.7.0
anyio==4.7.0
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
distro==1.9.0
Flask==3.1.0
Flask-Cors==5.0.0
fpdf==1.7.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.36.1
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.2
lazy_loader==0.4
MarkupSafe==3.0.2
networkx==3.4.2
numpy==2.1.3
openai==1.58.1
opencv-python==4.10.0.84
packaging==24.2
pillow==11.0.0
PyAudioWPatch~=0.2.12.6
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
pydub==0.25.1
PyPDF2==3.0.1
requests==2.32.3
scikit-image==0.24.0
scipy==1.14.1
sniffio==1.3.1
srt==3.5.3
tifffile==2024.9.20
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.2.3
vosk==0.3.45
websockets==14.1
Werkzeug==3.1.3
ollama~=0.4.7